Data packages are displayed with pagination to enable displaying silos with large number of data packages. For this, I use a mysql table to index the list of data packages in a silo to enable fast lookup. * List of silos displayed with more information * The path /{silo_name} and /{silo_name}/datasets use the same controller
Search for the three different types now works.
Admin controller and solr worker modified to broadcast silo creation and updates and index the silo metadata in solr.
Extra output fields can be passed to search using the parameter fl
File names are indexed with the path, to enable searching in a new solr field called filename
The search controller uses the filename field to search for files and directories

Modified sword submissions for adding new data packages to the table